move createRandomRoomName.js to /misc folder
