+++ src/components/launchers/RoomPickerOld.vue
@@ -135,7 +135,7 @@
 
 
 <script>
-import createRandomRoomName from "../createRandomRoomName";
+import createRandomRoomName from "/src/misc/createRandomRoomName.js";
 
 export default {
   name: 'app-roomPicker',
